chore: user 관련 dto, model 클래스를 하위 user 패키지로 이동
