+++ CatchMate/data/src/main/java/com/catchmate/data/datasource/remote/UserService.kt
@@ -1,13 +1,13 @@
 package com.catchmate.data.datasource.remote
 
-import com.catchmate.data.dto.GetUserProfileByIdResponseDTO
-import com.catchmate.data.dto.GetUserProfileResponseDTO
-import com.catchmate.data.dto.PatchUserAlarmResponseDTO
-import com.catchmate.data.dto.PatchUserProfileRequestDTO
-import com.catchmate.data.dto.PatchUserProfileResponseDTO
-import com.catchmate.data.dto.PostUserAdditionalInfoRequestDTO
-import com.catchmate.data.dto.PostUserAdditionalInfoResponseDTO
-import com.catchmate.domain.model.AlarmType
+import com.catchmate.data.dto.user.GetUserProfileByIdResponseDTO
+import com.catchmate.data.dto.user.GetUserProfileResponseDTO
+import com.catchmate.data.dto.user.PatchUserAlarmResponseDTO
+import com.catchmate.data.dto.user.PatchUserProfileRequestDTO
+import com.catchmate.data.dto.user.PatchUserProfileResponseDTO
+import com.catchmate.data.dto.user.PostUserAdditionalInfoRequestDTO
+import com.catchmate.data.dto.user.PostUserAdditionalInfoResponseDTO
+import com.catchmate.domain.model.enumclass.AlarmType
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
