+++ CatchMate/data/src/main/java/com/catchmate/data/dto/user/GetUserProfileResponseDTO.kt
@@ -1,4 +1,4 @@
-package com.catchmate.data.dto
+package com.catchmate.data.dto.user
 
 data class GetUserProfileResponseDTO(
     val userId: Long,
