+++ CatchMate/data/src/main/java/com/catchmate/data/dto/user/PatchUserProfileResponseDTO.kt
@@ -1,4 +1,4 @@
-package com.catchmate.data.dto
+package com.catchmate.data.dto.user
 
 data class PatchUserProfileResponseDTO(
     val state: Boolean,
