+++ CatchMate/data/src/main/java/com/catchmate/data/dto/user/PostUserAdditionalInfoResponseDTO.kt
@@ -1,4 +1,4 @@
-package com.catchmate.data.dto
+package com.catchmate.data.dto.user
 
 data class PostUserAdditionalInfoResponseDTO(
     val userId: Long,
