+++ CatchMate/domain/src/main/java/com/catchmate/domain/model/enumclass/AlarmType.kt
@@ -1,4 +1,4 @@
-package com.catchmate.domain.model
+package com.catchmate.domain.model.enumclass
 
 enum class AlarmType {
     ALL,
