+++ CatchMate/domain/src/main/java/com/catchmate/domain/model/user/GetUserProfileByIdResponse.kt
@@ -1,4 +1,4 @@
-package com.catchmate.domain.model
+package com.catchmate.domain.model.user
 
 data class GetUserProfileByIdResponse(
     val userId: Long,
