+++ CatchMate/domain/src/main/java/com/catchmate/domain/model/user/PatchUserAlarmResponse.kt
@@ -1,4 +1,6 @@
-package com.catchmate.domain.model
+package com.catchmate.domain.model.user
+
+import com.catchmate.domain.model.enumclass.AlarmType
 
 data class PatchUserAlarmResponse(
     val userId: Long,
