+++ CatchMate/domain/src/main/java/com/catchmate/domain/model/user/PostUserAdditionalInfoResponse.kt
@@ -1,4 +1,4 @@
-package com.catchmate.domain.model
+package com.catchmate.domain.model.user
 
 data class PostUserAdditionalInfoResponse(
     val userId: Long,
