+++ CatchMate/domain/src/main/java/com/catchmate/domain/usecase/user/GetUserProfileUseCase.kt
@@ -1,6 +1,6 @@
 package com.catchmate.domain.usecase.user
 
-import com.catchmate.domain.model.GetUserProfileResponse
+import com.catchmate.domain.model.user.GetUserProfileResponse
 import com.catchmate.domain.repository.UserRepository
 import javax.inject.Inject
