+++ CatchMate/presentation/src/main/java/com/catchmate/presentation/view/onboarding/CheerStyleOnboardingFragment.kt
@@ -9,7 +9,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
-import com.catchmate.domain.model.PostUserAdditionalInfoRequest
+import com.catchmate.domain.model.user.PostUserAdditionalInfoRequest
 import com.catchmate.presentation.R
 import com.catchmate.presentation.databinding.FragmentCheerStyleOnboardingBinding
 import com.catchmate.presentation.viewmodel.LocalDataViewModel
