+++ CatchMate/presentation/src/main/java/com/catchmate/presentation/viewmodel/HomeViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.catchmate.domain.exception.ReissueFailureException
 import com.catchmate.domain.model.GetBoardPagingResponse
-import com.catchmate.domain.model.GetUserProfileResponse
+import com.catchmate.domain.model.user.GetUserProfileResponse
 import com.catchmate.domain.usecase.board.GetBoardPagingUseCase
 import com.catchmate.domain.usecase.user.GetUserProfileUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
