+++ CatchMate/presentation/src/main/java/com/catchmate/presentation/viewmodel/SignUpViewModel.kt
@@ -6,8 +6,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.catchmate.domain.exception.ReissueFailureException
 import com.catchmate.domain.model.GetCheckNicknameResponse
-import com.catchmate.domain.model.PostUserAdditionalInfoRequest
-import com.catchmate.domain.model.PostUserAdditionalInfoResponse
+import com.catchmate.domain.model.user.PostUserAdditionalInfoRequest
+import com.catchmate.domain.model.user.PostUserAdditionalInfoResponse
 import com.catchmate.domain.usecase.auth.GetAuthCheckNicknameUseCase
 import com.catchmate.domain.usecase.user.PostUserAdditionalInfoUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
